+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export {AccommodationCard} from "./Cards/AccommodationCard"
-export {AccommodationSearch} from "./AccommodationSearch"
-export {AdvancedSearch} from "./AdvancedSearch"
+export {AccommodationSearch} from "./Search/AccommodationSearch"
+export {AdvancedSearch} from "./Search/AdvancedSearch"
 export {BookStayButton} from "./BookStayButton"
 export {CityCard} from "./Cards/CityCard"
 export {Footer} from "./Footer"
@@ -11,4 +11,4 @@ export {Rating} from "./Rating"
 export {ReservationCard} from "./Cards/ReservationCard"
 export {ReservationForm} from "./ReservationForm"
 export {SearchButton} from "./SearchButton"
-export {SimpleSearch} from "./SimpleSearch"
+export {SimpleSearch} from "./Search/SimpleSearch"
